Extract wave portfolio resolution helper
